// src/modules/omnia/interfaces/omnia-product.ts
export interface OmniaProduct {
  codfilial: string;
  codprod: number;
  descricao: string;
  unidade: string;
  embalagem: string;
  codepto: number;
  departamento: string;
  codsec: number;
  secao: string;
  codmarca: number;
  nomemarca: string;
  codbarra: number;
  codbarracaixa: string;
  pesoliq_gr: string;
  pesobruto_gr: string;
  altura_cm: string;
  largura_cm: string;
  comprimento_cm: string;
  ncm: string;
  informacoestecnicas: string;
  nomeecommerce: string;
  tituloecommerce: string;
  metatagdescricao: string;
  descricaolonga: string;
  descricaocurta: string;
  dtcadastro: string;
  dtultalter: string;
  multiplo: string;
  tipoestoque: string;
  pesovariavel: string;
  pesopeca: string;
  reqnum: number;
}